+++ index.js
@@ -3,12 +3,13 @@
  * @Author: wanglh
  * @LastEditors: wanglh
  * @Date: 2019-03-11 17:31:51
- * @LastEditTime: 2019-03-14 10:43:16
+ * @LastEditTime: 2019-03-14 15:39:30
  */
 
 import ButtonView from './src/ButtonView';
 
 import CornerTextView from './src/CornerTextView';
 
-export { ButtonView, CornerTextView };
-export default ButtonView;
+import LoadingView from './src/LoadingView';
+
+export { ButtonView, CornerTextView, LoadingView };