+++ CHANGELOG.md
@@ -1,2 +1,4 @@
+.[v1.0.3]  Add LoadingView, which provides display and hide methods
+.[v1.0.2]  Edit index.d.ts for CornerTextView
 .[v1.0.1]  Add CornerTextView, which can customize styles.
 .[v1.0.0]  Add ButtonView, which can customize styles.